refactor: Bookshelf setup 함수 분리
- Bookshelf setup함수 분리

  - item의 동작(Action)을 담당하는 setup

  - item의 Grid layout을 담당하는 setup

  2개의 setup을 이용하여 Bookshelf 구성
