+++ src/newtab/components/composition/setupBookshelfAction.ts
@@ -1,62 +1,35 @@
+import { OPEN_BOOKSHELF_MODALS } from "@/newtab/store/modules/bookshelfModal";
 import { ContextMenuTarget } from "@/newtab/store/modules/contextMenu";
 import {
   SET_TOOLTIP_POSITION,
   SET_TOOLTIP_SHOW,
   SET_TOOLTIP_TEXT,
 } from "@/newtab/store/modules/tooltip";
-import BookmarkApi from "@/newtab/utils/bookmarkApi";
 import { Item } from "@/shared/types/store";
-import { onMounted, computed, watch, ref, Ref } from "vue";
+import { Ref, SetupContext } from "vue";
 import { useStore } from "vuex";
-import { SET_REFRESH_TARGET } from "../../store/index";
 import { openContextMenu } from "../../utils/contextMenu";
 
-export interface SetupBookshelf {
+interface Props {
   folderItem: Ref<Item>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  context: SetupContext<Record<string, any>>;
+}
+
+export interface SetupBookshelf {
   openTooltip: (title: string, event: MouseEvent) => void;
   closeTooltip: () => void;
   openUrl: (id: string, url: string) => void;
   openContextMenu: (event: PointerEvent, target: ContextMenuTarget) => void;
+  onClickFolder: (item: Item) => void;
 }
-type Props = {
-  id: string;
-};
 
-export const setupBookshelf = (
-  props: Props,
-  onRefresh?: (folderItem: Item) => Promise<Item>
-): SetupBookshelf => {
+export const setupBookshelfAction = (props: Props): SetupBookshelf => {
+  const { folderItem, context } = props;
   const store = useStore();
-  const { id } = props;
 
   /**
-   * Data
-   */
-  const folderItem = ref({} as Item);
-  const refreshTargetId = computed(() => store.getters.GET_REFRESH_TARGET);
-
-  const refresh = async () => {
-    folderItem.value = await BookmarkApi.getSubTree(id);
-    if (onRefresh) {
-      folderItem.value = await onRefresh(folderItem.value);
-    }
-  };
-  watch(refreshTargetId, (newId) => {
-    if (newId && props.id === newId) {
-      refresh();
-      store.commit(SET_REFRESH_TARGET, "");
-    }
-  });
-
-  /**
-   * LifeCycle Hook
-   */
-  onMounted(() => {
-    refresh();
-  });
-
-  /**
-   * Method
+   * method
    */
   const openTooltip = (title: string, event: MouseEvent) => {
     const targetElement = event.target as HTMLElement;
@@ -73,19 +46,32 @@ export const setupBookshelf = (
     store.commit(SET_TOOLTIP_TEXT, title);
     store.commit(SET_TOOLTIP_SHOW, true);
   };
+
   const closeTooltip = () => {
     store.commit(SET_TOOLTIP_SHOW, false);
   };
+
   const openUrl = (id: string, url: string) => {
     window.open(url, "_blank")?.focus();
     store.commit(SET_TOOLTIP_SHOW, false);
   };
 
+  const onClickFolder = (item: Item): void => {
+    const isDesktop = folderItem.value.parentId === "0";
+    const { id, title } = item;
+    if (isDesktop) {
+      store.commit(OPEN_BOOKSHELF_MODALS, { id, title });
+    } else {
+      context.emit("routeInFolder", id);
+    }
+    store.commit(SET_TOOLTIP_SHOW, false);
+  };
+
   return {
-    folderItem,
     openTooltip,
     closeTooltip,
     openUrl,
     openContextMenu,
+    onClickFolder,
   };
 };
